improve_e2e_download_keys MXDownloadKeysPromisesManager becomes MXDeviceList to have the same behaviour than the JS implementation
